// src/api/generated/model/productDetail.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Shop API
 * RN 마이그레이션 검증용 API. 에러는 HTTP 상태가 아니라 `code`로 분기한다.
 * OpenAPI spec version: 0.1.0
 */

export interface ProductDetail {
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  id: number;
  title: string;
  price: number;
  discountPercentage: number;
  rating: number;
  thumbnail: string;
  brand?: string;
  description: string;
  category: string;
  images: string[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  stock: number;
  availabilityStatus?: string;
  shippingInformation?: string;
  warrantyInformation?: string;
  returnPolicy?: string;
}